+++ Utils/ClientExtensions.cs
@@ -4,6 +4,9 @@ namespace ChatTranslatorHud.Utils;
 
 public static class ClientExtensions
 {
-    public static bool IsValidPlayer(this IGameClient client) 
-        => client.IsValid && !client.IsFakeClient;
+    extension(IGameClient client)
+    {
+        public bool IsValidPlayer()
+            => client.IsValid && !client.IsFakeClient;
+    }
 }