+++ Utils/DictionaryExtensions.cs
@@ -4,10 +4,13 @@ namespace ChatTranslatorHud.Utils;
 
 public static class DictionaryExtensions
 {
-    public static void AddToLanguageGroup(this Dictionary<string, List<IGameClient>> groups, string lang, IGameClient client)
+    extension(Dictionary<string, List<IGameClient>> groups)
     {
-        if (!groups.TryGetValue(lang, out var list))
-            groups[lang] = list = [];
-        list.Add(client);
+        public void AddToLanguageGroup(string lang, IGameClient client)
+        {
+            if (!groups.TryGetValue(lang, out var list))
+                groups[lang] = list = [];
+            list.Add(client);
+        }
     }
 }